Flask BE/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
cachelib==0.9.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.6
cssselect2==0.7.0
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
packaging==22.0
passlib==1.7.4
Pillow==9.3.0
prompt-toolkit==3.0.36
pycparser==2.21
pydyf==0.5.0
pyparsing==3.0.9
pyphen==0.13.2
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
tinycss2==1.2.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
weasyprint==57.1
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
zopfli==0.2.2
